requirements.txt
# Core Dependencies
pandas>=1.5.0,<2.0.0
numpy>=1.24.0,<1.25.0
scikit-learn>=1.3.0,<1.4.0
joblib>=1.3.0,<1.4.0

# Web Framework & API
Flask>=2.2.0,<2.3.0
flask-cors>=4.0.0,<4.1.0
flask-limiter>=3.3.0,<3.4.0
requests>=2.31.0,<2.32.0

# Data Processing
scipy>=1.10.0,<1.11.0
pyarrow>=12.0.0,<13.0.0
python-dateutil>=2.8.0,<2.9.0
pytz>=2023.0,<2024.0

# Machine Learning
imbalanced-learn>=0.11.0,<0.12.0
threadpoolctl>=3.2.0,<3.3.0

# Visualization
plotly>=5.17.0,<5.18.0
matplotlib>=3.7.0,<3.8.0
seaborn>=0.12.0,<0.13.0

# Streamlit
streamlit>=1.27.0,<1.28.0
tenacity>=8.2.0,<8.3.0
toml>=0.10.0,<0.11.0
tornado>=6.3.0,<6.4.0
watchdog>=3.0.0,<3.1.0

# Image Processing
pillow>=9.5.0,<10.0.0

# Protocol & Utilities
protobuf>=4.24.0,<4.25.0
urllib3>=2.0.0,<2.1.0